+++ NEWS.md
@@ -12,6 +12,7 @@ title: "NEWS for the emmeans package"
   * `qdrg()` has replaced its `ordinal.dim` argument with `ordinal`, a list with
     elements `dim` and `mode` -- which now fully supports all the modes available
     for ordinal models (#444). (`ordinal.dim` still works for backward compatibility.)
+  * Fix to bookkeeping bug in `emtrends` (#448)
 
 ## emmeans 1.8.8
   * Bug correction in `contrast` when `tran` is a `list` (#428)